requirements.txt
pypdf
pillow
pdf2image
reportlab
PyMuPDF
pytesseract
cryptography
tkinter-tooltip
matplotlib
numpy
opencv-python